Add code from D253 Rig as of 3/6/2018 and a small file with notes on Spike2 scripting language
I have not looked through all of the code that is being added during this commit yet. I assume there will be some duplicates or outdated versions of
code that is already stored in the repo. The plan is to commit everything without touching it first, and then clean it up once it is all stored on the repo.