WIP - adding a pretest section for Trace's experiment
UI changes: Created a toolbar button to run a selection of tests

Other: Copied the code more or less from
Experiment_Alex_Nov2019_GagnonPredictor.s2s, so that we could
use the "order" and less of a state transition from
Experiment_Default.s2s.

Tech debt: Allowing runExpmt to be more than 0 or 1,
now it is set to 0, 1, or 2. In the future this should
be called something closer to "experimentSection", to make
it clear that runExpmt=2 isn't running it at 2x speed or something.