Updating waveform file paths, remove unused scripts
Several of these scripts are not used in other places, and
appear like they were temporary or testing ones.

The reason why we are keeping the
"Parameters_Univesal_temp_SumofSines.s2s" files
though is that it's being currently run in the experiments,
on the rig 241B. I am updating the filepath to the
expected waveform file, because we should store a complete copy
of the data acquisition code.