Separate D006 parameters from those on D241B
The parameter files used for D241B used to be modified from D006,
and it was confusing and not correct for us to merge the updates
specific to the D241B rig into the Parameters_D006.s2s file.

That file was also using the sequencer calibrated file
"Sequencer_D019Copy.pls" - which makes more sense to be renamed
to align with "Sequencer_D241B.pls" becauase the D019 was a room
in the basement and not the current rig setup for D241B.

I reverted the changes to "Parameters_D006.s2s".