Remove unused waveform file, undo changes to Experiment_Default
Most of the "play_waveform" file was commented out, and it's
unknown if there are current uses in the existing codebase.

Also removing the one-off comment from Experiment_Default.s2s,
because it's not relevant to the other experimenters.

The comment of using 1 instead of 2 for the prePostNum is in the
Experiment_Trace_Nov2022.s2s file, and is probably safe being stored
there.