Create simplified version of script for pre-test
After lots of struggling to work with the Spike2 code, and it's general
complexity and the setup, it seemed more reasonable to split it out into
another file based on how the configuration and other setups are.

In practice running this "pre-test" is just like running another
experiment, and the distinction between the two isn't super clear or
separate.

Issues I ran into incorporating it into a second "button" on the script:

1. Needing to change the "is the experiment running" state variable
   (runExmpt) to be 0, 1, or 2 instead of just 0 or 1. It felt strange to
   have it work that way
2. Configurations do not fit well - and adding another button adds to
   the confusion.
3. You can't call the pre-test function with multiple different variables easily,
   because of the way that ToolbarIdle%() run sequentially each time.
   The complexity of the logic required is high..we'd have to create
   multiple other states for the "runExpmt=0-10" or something, and some
   way to increment it?
4. Need to add additional variables and prevent people from running the
   normal experiment at the same time as the test experiment. I had
   disabled the button, but the UI experiment is limited - i.e. we can't
   tell the user why the button is disabled.