Updating the sign of the ChairResetDirection, DrumResetDirection
It looks like we actually need for the sign to be flipped.

There was a comment in the "Parameters_Universal" file saying
"  ' Scales *1 for D253, *-1 for D019" - and the D241B uses D019Copy
for the sequencer.

This prevents the weird flips at high velocities. I'm still trying to
understand the driftFix% though