Matlab updates for D241B
For some reason - the EyeTracking Matlab files are different enough on
the D241B rig, where the code on the "Master" Github branch is
incompatible with the data that is generated here.

When pressing on the "Dessacade & Get Scale Factors" on the "Eye
Analysis" piece, it brings up some errors.

Hypothesis:
* Channel file update? The importSpike in the getScaleFactors_App.m uses
  a different set of integers - likely this refers to the channels?