Only record data into arrays for calibration run
Before, experimentalists were experiencing time delays of up to
20 seconds when adjusting the cameras for the eyeCalibration.

This was due to a bug in the code, where we were writing to matrix
data when it was not preallocated, and we would get large
pauses at Matlab generated more room and space for the data.
For example, there would be long pauses at 4032 or so in the time arena,
because it would need to wait for Matlab to increase the array sizes.

Also removed an output statement and converted it into a print
statement, because it's generally best practices to do that way instead.

- Preallocate arrays for time vectors